Rename ExtensionPointComponent to ExtensionPointClassComponent
