+++ react/components/ExtensionPointClassComponent.tsx
